MAINT: Move concatenate to fastcall protocol
The refactor to use vectorcall/fastcall is obviously much better
if we don't have to go back and forth, for concatenate we get:

     arr = np.random.random(20)
     %timeit np.concatenate((arr, arr), axis=0)

Going from ~1.2µs to just below 1µs and all the way down to ~850ns
(fluctuates quite a lot down to 822 even).  ~40% speedup in total
which is not too shabby.